source/operators/requirements.txt
kopf
kubernetes
pykube-ng
pyyaml